+++ serialization/src/commonMain/kotlin/nl/adaptivity/xmlutil/serialization/DefaultFormatCache.kt
@@ -40,10 +40,16 @@ import kotlin.jvm.JvmStatic
  * intended to be stored on the config, thus reused through multiple serializations.
  * Note that this requires the `serialName` attribute of `SerialDescriptor` instances to be unique.
  */
-internal class DefaultFormatCache : FormatCache() {
+public class DefaultFormatCache : FormatCache() {
     private val typeDescCache = HashMap<TypeKey, XmlTypeDescriptor>()
     private val elemDescCache = HashMap<DescKey, XmlDescriptor>()
     private val pendingDescs = HashSet<DescKey>()
+
+    override fun copy(): DefaultFormatCache = DefaultFormatCache()
+
+    override fun unsafeCache(): DefaultFormatCache {
+        return this
+    }
 
     @OptIn(ExperimentalSerializationApi::class)
     override fun lookupType(namespace: Namespace?, serialDesc: SerialDescriptor, defaultValue: () -> XmlTypeDescriptor): XmlTypeDescriptor {
@@ -78,6 +84,7 @@ internal class DefaultFormatCache : FormatCache() {
     ): XmlDescriptor {
         val key = DescKey(overridenSerializer, serializerParent, tagParent.takeIf { it !== serializerParent }, canBeAttribute)
 
+        @OptIn(ExperimentalSerializationApi::class)
         check(pendingDescs.add(key)) {
             "Recursive lookup of ${serializerParent.elementSerialDescriptor.serialName}"
         }
@@ -109,9 +116,10 @@ internal class DefaultFormatCache : FormatCache() {
 
     private data class TypeKey(val namespace: String, val descriptor: SerialDescriptor)
 
-    companion object {
+    private companion object {
         @JvmStatic
         private fun TypeKey(namespace: String?, descriptor: SerialDescriptor) =
             DefaultFormatCache.TypeKey(namespace ?: "", descriptor)
     }
 }
+